# Conflicts:
#	src/main/java/com/teamtiger/productservice/products/models/GetProductDTO.java
#	src/main/java/com/teamtiger/productservice/products/services/ProductService.java

+++ src/main/java/com/teamtiger/productservice/products/models/GetProductDTO.java
@@ -7,7 +7,7 @@ public record GetProductDTO(
      UUID id,
      String name,
      double retailPrice,
-     double weight
+     int weight
 )
 {
 }
